+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- URL rewrites for /tyskysour and /novara-live to Novara Live category archive
+
+### Changed
+
+- TyskySour functionality cloned and renamed Novara Live (non breaking. TyskySour code to be removed after migration)
+
+### Depreciated
+
+- TyskySour archive template
+
 ## [3.8.1] - 2022-11-21
 
 ### Added